moved models into folder, fixed model id incrementing
