+++ server/db/models/Product.js
@@ -1,0 +1,31 @@
+const Sequelize = require('sequelize');
+const db = require('../db');
+//MODEL
+const Product = db.define('product', {
+  cost: {
+    type: Sequelize.FLOAT
+  },
+  description: {
+    type: Sequelize.TEXT
+  },
+  image: {
+    type: Sequelize.STRING
+  },
+  id: {
+  	defaultValue: Sequelize.UUIDV4,
+  	primaryKey: true,
+    type: Sequelize.UUID
+  },
+  inStock: {
+    type: Sequelize.BOOLEAN
+  },
+  name: {
+    type: Sequelize.STRING,
+    allowNull: false
+  },
+  stock: {
+    type: Sequelize.INTEGER
+  }
+});
+//EXPORT
+module.exports = Product;
