+++ server/db/models/Transaction.js
@@ -1,0 +1,13 @@
+const db = require('../db');
+const Sequelize = require('sequelize');
+//MODEL
+const Transaction = db.define('transaction', {
+  quantity: {
+    type: Sequelize.INTEGER
+  },
+  totalCost: {
+    type: Sequelize.FLOAT
+  }
+});
+//EXPORT
+module.exports = Transaction;
